requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
python-dotenv
google-api-python-client